Update Router and library contracts

+++ backend/v2-periphery/src/libraries/UniswapV2Library.sol
@@ -3,8 +3,18 @@ pragma solidity 0.8.23;
 
 import { IUniswapV2Pair } from "../../../v2-core/src/interfaces/IUniswapV2Pair.sol";
 
+/**
+ * @title UniswapV2Library
+ * @dev This library is used for various Uniswap V2 functionalities.
+ */
 library UniswapV2Library {
-    // returns sorted token addresses, used to handle return values from pairs sorted in this order
+    /**
+     * @dev Sorts the token addresses in ascending order.
+     * @param tokenA The first token address.
+     * @param tokenB The second token address.
+     * @return token0 The token address that is smaller.
+     * @return token1 The token address that is larger.
+     */
     function sortTokens(address tokenA, address tokenB) internal pure returns (address token0, address token1) {
         require(tokenA != tokenB, "UniswapV2Library: IDENTICAL_ADDRESSES");
 
@@ -13,24 +23,39 @@ library UniswapV2Library {
         require(token0 != address(0), "UniswapV2Library: ZERO_ADDRESS");
     }
 
-    // calculates the CREATE2 address for a pair without making any external calls
+    /**
+     * @dev Calculates the CREATE2 address for a pair without making any external calls.
+     * @param factory The factory contract address.
+     * @param tokenA The first token address.
+     * @param tokenB The second token address.
+     * @return pair The pair address.
+     */
     function pairFor(address factory, address tokenA, address tokenB) internal pure returns (address pair) {
         (address token0, address token1) = sortTokens(tokenA, tokenB);
         pair = address(
-            uint256(
-                keccak256(
-                    abi.encodePacked(
-                        hex"ff",
-                        factory,
-                        keccak256(abi.encodePacked(token0, token1)),
-                        hex"96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f" // init code hash
+            uint160(
+                uint256(
+                    keccak256(
+                        abi.encodePacked(
+                            hex"ff",
+                            factory,
+                            keccak256(abi.encodePacked(token0, token1)),
+                            hex"96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f" // init code hash
+                        )
                     )
                 )
             )
         );
     }
 
-    // fetches and sorts the reserves for a pair
+    /**
+     * @dev Fetches and sorts the reserves for a pair.
+     * @param factory The factory contract address.
+     * @param tokenA The first token address.
+     * @param tokenB The second token address.
+     * @return reserveA The reserve of tokenA.
+     * @return reserveB The reserve of tokenB.
+     */
     function getReserves(address factory, address tokenA, address tokenB)
         internal
         view
@@ -41,7 +66,13 @@ library UniswapV2Library {
         (reserveA, reserveB) = tokenA == token0 ? (reserve0, reserve1) : (reserve1, reserve0);
     }
 
-    // given some amount of an asset and pair reserves, returns an equivalent amount of the other asset
+    /**
+     * @dev Given some amount of an asset and pair reserves, returns an equivalent amount of the other asset.
+     * @param amountA The amount of the first asset.
+     * @param reserveA The reserve of the first asset.
+     * @param reserveB The reserve of the second asset.
+     * @return amountB The equivalent amount of the second asset.
+     */
     function quote(uint256 amountA, uint256 reserveA, uint256 reserveB) internal pure returns (uint256 amountB) {
         require(amountA > 0, "UniswapV2Library: INSUFFICIENT_AMOUNT");
         require(reserveA > 0 && reserveB > 0, "UniswapV2Library: INSUFFICIENT_LIQUIDITY");
@@ -49,7 +80,13 @@ library UniswapV2Library {
         amountB = (amountA * reserveB) / reserveA;
     }
 
-    // given an input amount of an asset and pair reserves, returns the maximum output amount of the other asset
+    /**
+     * @dev Given an input amount of an asset and pair reserves, returns the maximum output amount of the other asset.
+     * @param amountIn The input amount of the first asset.
+     * @param reserveIn The reserve of the first asset.
+     * @param reserveOut The reserve of the second asset.
+     * @return amountOut The maximum output amount of the second asset.
+     */
     function getAmountOut(uint256 amountIn, uint256 reserveIn, uint256 reserveOut)
         internal
         pure
@@ -64,7 +101,13 @@ library UniswapV2Library {
         amountOut = numerator / denominator;
     }
 
-    // given an output amount of an asset and pair reserves, returns a required input amount of the other asset
+    /**
+     * @dev Given an output amount of an asset and pair reserves, returns a required input amount of the other asset.
+     * @param amountOut The output amount of the first asset.
+     * @param reserveIn The reserve of the first asset.
+     * @param reserveOut The reserve of the second asset.
+     * @return amountIn The required input amount of the second asset.
+     */
     function getAmountIn(uint256 amountOut, uint256 reserveIn, uint256 reserveOut)
         internal
         pure
@@ -78,7 +121,13 @@ library UniswapV2Library {
         amountIn = (numerator / denominator) + 1;
     }
 
-    // performs chained getAmountOut calculations on any number of pairs
+    /**
+     * @dev Performs chained getAmountOut calculations on any number of pairs.
+     * @param factory The factory contract address.
+     * @param amountIn The input amount of the first asset.
+     * @param path The path of pairs to swap along.
+     * @return amounts The output amounts of the other assets.
+     */
     function getAmountsOut(address factory, uint256 amountIn, address[] memory path)
         internal
         view
@@ -95,7 +144,13 @@ library UniswapV2Library {
         }
     }
 
-    // performs chained getAmountIn calculations on any number of pairs
+    /**
+     * @dev Performs chained getAmountIn calculations on any number of pairs.
+     * @param factory The factory contract address.
+     * @param amountOut The output amount of the first asset.
+     * @param path The path of pairs to swap along.
+     * @return amounts The required input amounts of the other assets.
+     */
     function getAmountsIn(address factory, uint256 amountOut, address[] memory path)
         internal
         view
